pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gadash"
version = "0.1.0"
description = "GA optimization + dashboard"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.23",
  "pyyaml>=6.0",
  "pillow>=10.0",
  "fastapi>=0.110",
  "uvicorn>=0.23",
  "torch",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
